Append / if not provided in baseURL
Fixes 446 by making sure that the baseURL of an AFHTTPClient is always
a directory URL rather than a file URL.

We assume the URL is intended to be a directory URL, and append / if
necessary so that later use of the baseURL does not strip any of the
path of the URL.

I think fixing this makes more sense than documenting it, because:

- passing a file baseURL is always be a mistake
- the check and fix is small

This may not be the right fix; you may prefer to return nil or even
fail an assertion in this case. I don't know if you favour forcing
people to fix their code vs. adapting to their intent. But I think not
recognizing this as an error should be fixed.